python_app/requirements.txt
psycopg2-binary
pandas
requests
dbt-postgres
